Refactored Divisions to remove duplicate code and fixed rework of GBD
